Solved Mapping Errors

+++ teatru-reservations/src/main/java/com/example/teatru_reservations/repository/ActorRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ActorRepository extends JpaRepository<Actor, Long> {
+public interface ActorRepository extends JpaRepository<Actor, Integer> {
     Optional<Actor> findByActorName(String actorName);
 
     Optional<Actor> findById(Integer id);
